Mower Inventory Project
A project that reads data from MP2_Dataset.txt, and outputs it based on the type of mower the user wants to buy, through a while loop in the client class. The data for each mower is reformatted and saved in test.txt, and additional options could include customizing the PrintWriter's output so that the mowers are organized by subclass. Abstract methods are heavily implemented for similar objects, with list methods being used to interpret the input data and make sure classes requiring various parameters are handled successfully.